move get api call function
